New feature category -> single article
I've cleaned up code and added a new feature.

When you have a category view with articles that do not have menu items and you view a single article. It now takes the article title and article intro-text as title and description in your masthead. It will also check if there is a image_intro or image_fulltext available and use that instead. If params are not present it will fall back to menu item specific masthead or default masthead.